requirements.txt
black
colorama
mkdocs
matplotlib
Pillow
tensorflow
scipy
opencv-python
